requirements.txt
flask
flask-sqlalchemy
python-dotenv
html2image
gunicorn==21.2.0
